requirements.txt
anyio==4.2.0
asgiref==3.7.2
channels==4.0.0
click==8.1.7
Django==4.2.9
djangorestframework==3.14.0
exceptiongroup==1.2.0
h11==0.14.0
httptools==0.6.1
idna==3.6
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
uvicorn==0.27.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
